requirements.txt
Flask>=2.0
elasticsearch>=8.0,<9.0
requests>=2.25
python-dotenv>=0.19